microservices/microservice_parsing/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --extra-index-url=https://download.pytorch.org/whl/cpu requirements.in
#
--extra-index-url https://download.pytorch.org/whl/cpu

accelerate==1.6.0
    # via -r requirements.in
albumentations==1.4.2
    # via vietocr
annotated-types==0.7.0
    # via pydantic
anyascii==0.3.2
    # via python-doctr
anyio==4.9.0
    # via
    #   httpx
    #   mcp
    #   openai
    #   sse-starlette
    #   starlette
beautifulsoup4==4.13.4
    # via gdown
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via
    #   typer
    #   uvicorn
contourpy==1.3.2
    # via matplotlib
cycler==0.12.1
    # via matplotlib
defusedxml==0.7.1
    # via python-doctr
distro==1.9.0
    # via openai
easyocr==1.7.2
    # via -r requirements.in
einops==0.2.0
    # via vietocr
et-xmlfile==2.0.0
    # via openpyxl
fastapi==0.115.12
    # via -r requirements.in
filelock==3.18.0
    # via
    #   gdown
    #   huggingface-hub
    #   torch
    #   transformers
fonttools==4.57.0
    # via matplotlib
fsspec==2025.3.2
    # via
    #   huggingface-hub
    #   torch
gdown==4.4.0
    # via vietocr
grpcio==1.67.1
    # via pymilvus
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
h5py==3.13.0
    # via python-doctr
httpcore==1.0.8
    # via httpx
httpx==0.28.1
    # via
    #   -r requirements.in
    #   mcp
    #   openai
httpx-sse==0.4.0
    # via mcp
huggingface-hub==0.30.2
    # via
    #   accelerate
    #   python-doctr
    #   tokenizers
    #   transformers
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
imageio==2.37.0
    # via
    #   imgaug
    #   scikit-image
imgaug==0.4.0
    # via vietocr
jinja2==3.1.6
    # via torch
jiter==0.9.0
    # via openai
joblib==1.4.2
    # via scikit-learn
kiwisolver==1.4.8
    # via matplotlib
langdetect==1.0.9
    # via python-doctr
langid==1.1.6
    # via -r requirements.in
lazy-loader==0.4
    # via scikit-image
lmdb==1.6.2
    # via vietocr
loguru==0.7.3
    # via -r requirements.in
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
matplotlib==3.10.1
    # via imgaug
mcp[cli]==1.6.0
    # via -r requirements.in
mdurl==0.1.2
    # via markdown-it-py
milvus-lite==2.4.12
    # via pymilvus
mpmath==1.3.0
    # via sympy
networkx==3.4.2
    # via
    #   scikit-image
    #   torch
ninja==1.11.1.4
    # via easyocr
numpy==2.2.5
    # via
    #   accelerate
    #   albumentations
    #   contourpy
    #   easyocr
    #   h5py
    #   imageio
    #   imgaug
    #   langid
    #   matplotlib
    #   opencv-python
    #   opencv-python-headless
    #   pandas
    #   python-doctr
    #   scikit-image
    #   scikit-learn
    #   scipy
    #   shapely
    #   tifffile
    #   torchvision
    #   transformers
openai==1.76.0
    # via -r requirements.in
opencv-python==4.11.0.86
    # via
    #   imgaug
    #   python-doctr
opencv-python-headless==4.11.0.86
    # via
    #   -r requirements.in
    #   albumentations
    #   easyocr
openpyxl==3.1.5
    # via -r requirements.in
packaging==25.0
    # via
    #   accelerate
    #   huggingface-hub
    #   lazy-loader
    #   matplotlib
    #   scikit-image
    #   transformers
pandas==2.2.3
    # via
    #   -r requirements.in
    #   pymilvus
pillow==10.2.0
    # via
    #   easyocr
    #   imageio
    #   imgaug
    #   matplotlib
    #   python-doctr
    #   scikit-image
    #   torchvision
    #   vietocr
pip-chill==1.0.3
    # via -r requirements.in
prefetch-generator==1.0.1
    # via vietocr
protobuf==6.30.2
    # via pymilvus
psutil==7.0.0
    # via accelerate
pyclipper==1.3.0.post6
    # via
    #   easyocr
    #   python-doctr
pydantic==2.11.3
    # via
    #   -r requirements.in
    #   fastapi
    #   mcp
    #   openai
    #   pydantic-settings
pydantic-core==2.33.1
    # via pydantic
pydantic-settings==2.9.1
    # via
    #   -r requirements.in
    #   mcp
pygments==2.19.1
    # via rich
pymilvus==2.5.7
    # via -r requirements.in
pymupdf==1.25.5
    # via -r requirements.in
pyparsing==3.2.3
    # via matplotlib
pypdfium2==4.30.1
    # via python-doctr
pysocks==1.7.1
    # via requests
python-bidi==0.6.6
    # via easyocr
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
python-doctr==0.11.0
    # via -r requirements.in
python-dotenv==1.1.0
    # via
    #   mcp
    #   pydantic-settings
    #   pymilvus
python-multipart==0.0.20
    # via -r requirements.in
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via
    #   accelerate
    #   albumentations
    #   easyocr
    #   huggingface-hub
    #   transformers
rapidfuzz==3.13.0
    # via python-doctr
regex==2024.11.6
    # via
    #   tiktoken
    #   transformers
requests[socks]==2.32.3
    # via
    #   -r requirements.in
    #   gdown
    #   huggingface-hub
    #   tiktoken
    #   transformers
rich==14.0.0
    # via typer
safetensors==0.5.3
    # via
    #   accelerate
    #   transformers
scikit-image==0.25.2
    # via
    #   albumentations
    #   easyocr
    #   imgaug
    #   vietocr
scikit-learn==1.6.1
    # via albumentations
scipy==1.15.2
    # via
    #   -r requirements.in
    #   albumentations
    #   easyocr
    #   imgaug
    #   python-doctr
    #   scikit-image
    #   scikit-learn
shapely==2.1.0
    # via
    #   easyocr
    #   imgaug
    #   python-doctr
shellingham==1.5.4
    # via typer
six==1.17.0
    # via
    #   gdown
    #   imgaug
    #   langdetect
    #   python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   openai
soupsieve==2.7
    # via beautifulsoup4
sse-starlette==2.3.3
    # via mcp
starlette==0.46.2
    # via
    #   fastapi
    #   mcp
    #   sse-starlette
sympy==1.13.3
    # via torch
threadpoolctl==3.6.0
    # via scikit-learn
tifffile==2025.3.30
    # via scikit-image
tiktoken==0.9.0
    # via -r requirements.in
tokenizers==0.21.1
    # via transformers
torch==2.7.0+cpu
    # via
    #   -r requirements.in
    #   accelerate
    #   easyocr
    #   torchvision
torchvision==0.22.0+cpu
    # via
    #   -r requirements.in
    #   easyocr
tqdm==4.67.1
    # via
    #   gdown
    #   huggingface-hub
    #   milvus-lite
    #   openai
    #   python-doctr
    #   transformers
transformers==4.51.3
    # via -r requirements.in
typer==0.15.2
    # via mcp
typing-extensions==4.13.2
    # via
    #   albumentations
    #   anyio
    #   beautifulsoup4
    #   fastapi
    #   huggingface-hub
    #   openai
    #   pydantic
    #   pydantic-core
    #   torch
    #   typer
    #   typing-inspection
typing-inspection==0.4.0
    # via
    #   pydantic
    #   pydantic-settings
tzdata==2025.2
    # via pandas
ujson==5.10.0
    # via pymilvus
urllib3==2.4.0
    # via requests
uvicorn==0.34.2
    # via
    #   -r requirements.in
    #   mcp
vietocr==0.3.13
    # via -r requirements.in
websockets==15.0.1
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
